Move manifest_populator into spark_job/validation
